modularize surface_load appropriately
