+++ src/engine/surface_load_objects.c
@@ -1,0 +1,170 @@
+#include <PR/ultratypes.h>
+
+#include "sm64.h"
+#include "game/ingame_menu.h"
+#include "graph_node.h"
+#include "behavior_script.h"
+#include "behavior_data.h"
+#include "game/memory.h"
+#include "game/object_helpers.h"
+#include "game/macro_special_objects.h"
+#include "surface_collision.h"
+#include "game/mario.h"
+#include "game/object_list_processor.h"
+#include "surface_load.h"
+
+/**
+ * Read the data for vertices for reference by triangles.
+ */
+static s16 *read_vertex_data(s16 **data) {
+    s32 numVertices;
+    UNUSED u8 filler[16];
+    s16 *vertexData;
+
+    numVertices = *(*data);
+    (*data)++;
+
+    vertexData = *data;
+    *data += 3 * numVertices;
+
+    return vertexData;
+}
+
+/**
+ * Applies an object's transformation to the object's vertices.
+ */
+void transform_object_vertices(s16 **data, s16 *vertexData) {
+    register s16 *vertices;
+    register f32 vx, vy, vz;
+    register s32 numVertices;
+
+    Mat4 *objectTransform;
+    Mat4 m;
+
+    objectTransform = &gCurrentObject->transform;
+
+    numVertices = *(*data);
+    (*data)++;
+
+    vertices = *data;
+
+    if (gCurrentObject->header.gfx.throwMatrix == NULL) {
+        gCurrentObject->header.gfx.throwMatrix = objectTransform;
+        obj_build_transform_from_pos_and_angle(gCurrentObject, O_POS_INDEX, O_FACE_ANGLE_INDEX);
+    }
+
+    obj_apply_scale_to_matrix(gCurrentObject, m, *objectTransform);
+
+    // Go through all vertices, rotating and translating them to transform the object.
+    while (numVertices--) {
+        vx = *(vertices++);
+        vy = *(vertices++);
+        vz = *(vertices++);
+
+        //! No bounds check on vertex data
+        *vertexData++ = (s16)(vx * m[0][0] + vy * m[1][0] + vz * m[2][0] + m[3][0]);
+        *vertexData++ = (s16)(vx * m[0][1] + vy * m[1][1] + vz * m[2][1] + m[3][1]);
+        *vertexData++ = (s16)(vx * m[0][2] + vy * m[1][2] + vz * m[2][2] + m[3][2]);
+    }
+
+    *data = vertices;
+}
+
+/**
+ * Load in the surfaces for the gCurrentObject. This includes setting the flags, exertion, and room.
+ */
+void load_object_surfaces(s16 **data, s16 *vertexData) {
+    s32 surfaceType;
+    s32 i;
+    s32 numSurfaces;
+    s16 hasForce;
+    s16 flags;
+    s16 room;
+
+    surfaceType = *(*data);
+    (*data)++;
+
+    numSurfaces = *(*data);
+    (*data)++;
+
+    hasForce = surface_has_force(surfaceType);
+
+    flags = surf_has_no_cam_collision(surfaceType);
+    flags |= SURFACE_FLAG_DYNAMIC;
+
+    // The DDD warp is initially loaded at the origin and moved to the proper
+    // position in paintings.c and doesn't update its room, so set it here.
+    
+    if (gCurrentObject->behavior == segmented_to_virtual(bhvDddWarp)) {
+        room = 5;
+    } else {
+        room = 0;
+    }
+
+    for (i = 0; i < numSurfaces; i++) {
+        struct Surface *surface = read_surface_data(vertexData, data);
+
+        if (surface != NULL) {
+            surface->object = gCurrentObject;
+            surface->type = surfaceType;
+
+            if (hasForce) {
+                surface->force = *(*data + 3);
+            } else {
+                surface->force = 0;
+            }
+
+            surface->flags |= flags;
+            surface->room = (s8) room;
+            add_surface(surface, TRUE);
+        }
+
+        if (hasForce) {
+            *data += 4;
+        } else {
+            *data += 3;
+        }
+    }
+}
+
+/**
+ * Transform an object's vertices, reload them, and render the object.
+ */
+void load_object_collision_model(void) {
+    UNUSED u8 filler[4];
+    s16 vertexData[600];
+
+    s16 *collisionData = gCurrentObject->collisionData;
+    f32 marioDist = gCurrentObject->oDistanceToMario;
+    f32 tangibleDist = gCurrentObject->oCollisionDistance;
+
+    // On an object's first frame, the distance is set to 19000.0f.
+    // If the distance hasn't been updated, update it now.
+    if (gCurrentObject->oDistanceToMario == 19000.0f) {
+        marioDist = dist_between_objects(gCurrentObject, gMarioObject);
+    }
+
+    // If the object collision is supposed to be loaded more than the
+    // drawing distance of 4000, extend the drawing range.
+    if (gCurrentObject->oCollisionDistance > 4000.0f) {
+        gCurrentObject->oDrawingDistance = gCurrentObject->oCollisionDistance;
+    }
+
+    // Update if no Time Stop, in range, and in the current room.
+    if (!(gTimeStopState & TIME_STOP_ACTIVE) && marioDist < tangibleDist
+        && !(gCurrentObject->activeFlags & ACTIVE_FLAG_IN_DIFFERENT_ROOM)) {
+        collisionData++;
+        transform_object_vertices(&collisionData, vertexData);
+
+        // TERRAIN_LOAD_CONTINUE acts as an "end" to the terrain data.
+        while (*collisionData != TERRAIN_LOAD_CONTINUE) {
+            load_object_surfaces(&collisionData, vertexData);
+        }
+    }
+
+    if (marioDist < gCurrentObject->oDrawingDistance) {
+        gCurrentObject->header.gfx.node.flags |= GRAPH_RENDER_ACTIVE;
+    } else {
+        gCurrentObject->header.gfx.node.flags &= ~GRAPH_RENDER_ACTIVE;
+    }
+}
