+++ src/game/object_list_processor.c
@@ -121,22 +121,22 @@ s16 gPrevFrameObjectCount;
  * The total number of surface nodes allocated (a node is allocated for each
  * spatial partition cell that a surface intersects).
  */
-s32 gSurfaceNodesAllocated;
+extern s32 gSurfaceNodesAllocated;
 
 /**
  * The total number of surfaces allocated.
  */
-s32 gSurfacesAllocated;
+extern s32 gSurfacesAllocated;
 
 /**
  * The number of nodes that have been created for surfaces.
  */
-s32 gNumStaticSurfaceNodes;
+extern s32 gNumStaticSurfaceNodes;
 
 /**
  * The number of surfaces in the pool.
  */
-s32 gNumStaticSurfaces;
+extern s32 gNumStaticSurfaces;
 
 /**
  * A pool used by chain chomp and wiggler to allocate their body parts.
